requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
supabase==2.0.2
openai==1.3.7
cassandra-driver==3.28.0
python-multipart==0.0.6
psutil 